+++ myexaminer/src/main/java/com/myexaminer/repository/ExerciseRepository.java
@@ -7,8 +7,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ExerciseRepository extends JpaRepository<Exercise, Integer> {
-    long deleteByIdExercise(int idExercise);
-    Optional<Exercise> findByIdExercise(int idExercise);
+public interface ExerciseRepository extends JpaRepository<Exercise, Long> {
+    long deleteByIdExercise(Long idExercise);
+    Optional<Exercise> findByIdExercise(Long idExercise);
     List<Exercise> findByExam(Exam exam);
 }
